move base64 into a its own files
