+++ readme.md
@@ -32,8 +32,17 @@ To run:
 ```
 LoRaMAC-node/build$ src/apps/LoRaMac/LoRaMac-classA
 ```
+
+To test:
+
+JoinResponse:
 ```
 {"txpk":{"imme":true, "freq":861.3, "rfch":0, "power":12, "modu":"FSK", "datr":50000, "fdev":3000, "size":17, "data":"IGMwaRINjbqvd4TSfHId/j0="}}
+```
+
+ConfirmedUp
+```
+{"txpk":{"data":"YAAAAAIKAAADMAAAcAMw/wAwtGyA8Q==","datr":[83,70,55,66,87,53,48,48],"freq":924.5,"size":22}}
 ```
 
 ## Introduction
