+++ src/radio/radio-simul/base64.h
@@ -1,0 +1,15 @@
+#ifndef base64_h
+#define base64_h
+
+#include <stddef.h>
+#include <stdint.h>
+
+char *base64_encode(const unsigned char *data,
+                    size_t input_length,
+                    size_t *output_length);
+
+unsigned char *base64_decode(const char *data,
+                             size_t input_length,
+                             size_t *output_length);
+
+#endif
